+++ src/ui/common/src/index.tsx
@@ -16,6 +16,7 @@ import { SnowflakeCard } from './components/integrations/cards/snowflakeCard';
 import { SqlServerCard } from './components/integrations/cards/sqlServerCard';
 import { ConnectedIntegrations } from './components/integrations/connectedIntegrations';
 import AddTableDialog from './components/integrations/dialogs/addTableDialog';
+import DeleteIntegrationDialog from './components/integrations/dialogs/deleteIntegrationDialog';
 import { BigQueryDialog } from './components/integrations/dialogs/bigqueryDialog';
 import { CSVDialog } from './components/integrations/dialogs/csvDialog';
 import { IntegrationDialog } from './components/integrations/dialogs/dialog';
@@ -274,6 +275,7 @@ export {
   AccountPage,
   AddIntegrations,
   addTable,
+  DeleteIntegrationDialog,
   AddTableDialog,
   AllTransition,
   AqueductBezier,
